[feat]: create SiteUser Entity

+++ src/main/java/com/ll/exam/sbb/answer/AnswerController.java
@@ -22,7 +22,7 @@ public class AnswerController {
 
     @PostMapping("/create/{id}")
 
-    public String detail(Model model, @PathVariable int id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
+    public String detail(Model model, @PathVariable long id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
         Question question = this.questionService.getQuestion(id);
 
         if ( bindingResult.hasErrors() ) {
