+++ src/main/java/com/ll/exam/sbb/question/QuestionController.java
@@ -28,7 +28,7 @@ public class QuestionController {
     }
 
     @GetMapping("/detail/{id}")
-    public String detail(Model model, @PathVariable int id, AnswerForm answerForm)  {
+    public String detail(Model model, @PathVariable long id, AnswerForm answerForm)  {
         Question question = questionService.getQuestion(id);
 
         model.addAttribute("question", question);
