+++ src/main/java/com/ll/exam/sbb/question/QuestionService.java
@@ -25,7 +25,7 @@ public class QuestionService {
         return this.questionRepository.findAll(pageable);
     }
 
-    public Question getQuestion(int id) {
+    public Question getQuestion(long id) {
         return questionRepository.findById(id)
                 //orElseThrow 적용
                 .orElseThrow(() -> new DataNotFoundException("no %d question not found,".formatted(id)));
